Major refactoring changes:
-ALL scene data structure objects now (should be) writable.
Switched over from arrays to lists, e.g. faces have a List<int> rather than int[].
All lists are guaranteed to exist, but are empty at object creation.

-ALL scene data structures now implement IMarshalable and have public constructors. IMarshalable represents
a class that can be marshaled to and from unmanaged memory. The entire marshaling mechanism has been re-written
to go through this interface, providing a streamlined interop capability that works in both directions

-Added InternalInterop stubs that get function calls inlined/method bodies replaced with injected
IL for fast interop (based on the SharpDX SharpCLI generator, using the generator for Tesla 3D Engine)

-Fleshed out MemoryHelper considerable to employ new fast interop functionality, added new ToNative/FromNative
helper methods for working with IMarshalable interfaces

-Removed scene validation in AssimpImporter to allow for scenes that only hold animations

-Spruced up MaterialProperty to better reflect how it is natively handled. All values are now read/written to the byte
buffer, using the fast-interop memory helper methods

-Spruced up Material with how textures are handled, and it works more like how aiMaterial works in the native library.

-Added many new properties to Material dealing with textures

-Redid embedded textures to work better with new IMarshalable techniques. Rather than splitting it up into compressed/non-compressed classes,
it is a single class now that reflects what aiTexture is in the native library.

-New marshaling to native functionality not yet fully adopted, the old mechanism (interl Ctors) is still
in use until further testing. The scene class has To/From/Free unamanaged scene methods that can be used to marshal
the managed scene data structure. Since there is no AssimpExporter yet, you have to use the AssimpLibrary export methods.
So far it seems to work OK, and the unmanaged data seems to be getting freed OK, but this is a big change.

-Made entire public library CLS compliant (removed all uint references, added CLS Compliant (false) attributes where appropiate).
Added a method in Mesh to get uint indices if desired.